+++ HISTORY.rst
@@ -6,6 +6,11 @@ Release History
 Unreleased
 ++++++++++
 
+**Data Migration**
+
+* ALCYN-1568: Fix delivered and invoiced quantities on partially delivered and expired sale orders.
+* ALCYN-1568: Set canceled quantities on partially delivered and expired sale orders.
+
 **Features and Improvements**
 
 **Bugfixes**